Abstract out reward image choosing/removing logic into its own component(RewardImageCard) on CreateRewards page
